robot-name: Changed test structure
